Replaced BGP parser and made store usable
Replaced the BGP parser from bmpreceiver with the one from pybgp which
is a bit more flexible and better suited for this task. Unfortunately it
did not support 32 bit AS numbers, so to use it together with bmpd this
needs to be patched. A simple modification is implemented, It'll be
distributed somehow later on...

The store was improved and now actually stores data. As the BGP path
attribute 'origin' is needed for route selection, also this is now
logged with each row in adj_rib_in.