Added daemonizing-support
bmpd can now be run as a daemon (\o/) with help from the python-daemon
package (http://pypi.python.org/pypi/python-daemon/).
Fixes #5.

Also moved creation of storage class to the exeutable (bmpd), as we need
access to this object for graceful shutdown of database processes. This
also made it easier to make number of database processe configurable as
a command-line argument, which it now is!
Refs #6.